Print.cpp was completely reconstructed. Memory & stack usage was reduced. Added special methods to prevent use of float type 'just for printing'. Added method fixPrint() that use int or long and number of precision to print the fixed point numbers right way. Added method dumpPrint() that helps to dump some debug data in hex format to specified output. Added method begin_() to HardwareSerial. Use of this reduces stack usage & memory consumption. It uses native ZM5101 format of baudrate: word value in 100s of bits like 1152 instead of 115200. Library ZUNO_OLED was reconstructed in same way to prevent stack overflow.

+++ hardware/arduino/zuno/cores/zuno/Stream.cpp
@@ -204,7 +204,7 @@ size_t Stream::readBytes(char *buffer, size_t length)
 {
   size_t count = 0;
   while (count < length) {
-    int c = timedRead();
+    uint8_t c = timedRead();
     if (c < 0) break;
     *buffer++ = (char)c;
     count++;
@@ -222,7 +222,7 @@ size_t Stream::readBytesUntil(char terminator, char *buffer, size_t length)
   if (length < 1) return 0;
   size_t index = 0;
   while (index < length) {
-    int c = timedRead();
+    uint8_t c = timedRead();
     if (c < 0 || c == terminator) break;
     *buffer++ = (char)c;
     index++;
